Stormwater Interface connects on ROSE\issacg.  Solution for object not int problem: change from/to scalar and back again.

+++ DotNet/StormwaterInterface/SWI_2/SWI_2/FormAddView.cs
@@ -60,7 +60,7 @@ namespace SWI_2
                 taV.Insert((int)comboBoxSubwatershed.SelectedValue, (int)numericUpDownAddView.Value, "");
                 
                 SWI_2.SANDBOXDataSetTableAdapters.SWSP_SURVEY_PAGETableAdapter taS = new SWI_2.SANDBOXDataSetTableAdapters.SWSP_SURVEY_PAGETableAdapter();
-                taS.Insert(taV.ScalarQuery(), (int)numericUpDownAddFirstSurveyPage.Value, System.DateTime.Today, "", "");
+                taS.Insert((int)taV.ScalarQuery(), (int)numericUpDownAddFirstSurveyPage.Value, System.DateTime.Today, "", "");
                 //_MyParentForm is for base users, so they need to be told what they were doing.
                 //homefully administrators can remember.
                 if (_MyParentForm != null)
